Consolidate browser tasks
